move "fixing december 1985 states"
after some thought, I decided that it makes sense to change the states
from BR to the right thing BEFORE merging with person data.

the reason is that while the state is part of the key for merging
persons/hh, we don't in fact use that key before 1990 because of the
household-person-person data structure. (Instead, we merge with
`n_entry`).